requirements.txt
python-dotenv
paho-mqtt
mysql-connector-python
streamlit
streamlit-extras
great-tables
numpy
wave
psycopg2-binary
pydub
Flask
Flask-Cors
mysql-connector-python
pandas
python-dotenv
mysql